refactor: create `ApiClient` out of check.py and requests_to_backend.py
