+++ samanthas_telegram_bot/conversation/callbacks/registration/student.py
@@ -3,7 +3,7 @@ import logging
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
 from telegram.constants import ParseMode
 
-from samanthas_telegram_bot.api_queries.check import get_level_of_written_test
+from samanthas_telegram_bot.api_queries.api_client import ApiClient
 from samanthas_telegram_bot.api_queries.smalltalk import send_user_data_get_smalltalk_test
 from samanthas_telegram_bot.conversation.auxil.callback_query_reply_sender import (
     CallbackQueryReplySender as CQReplySender,
@@ -95,8 +95,8 @@ async def assessment_store_answer_ask_question(
         len(context.user_data.student_assessment_answers)
         == len(context.chat_data.assessment.questions)
     ) or data == CommonCallbackData.ABORT:
-        context.user_data.student_assessment_resulting_level = await get_level_of_written_test(
-            context=context,
+        context.user_data.student_assessment_resulting_level = (
+            await ApiClient.get_level_of_written_test(context=context)
         )
         if context.user_data.student_assessment_resulting_level in LEVELS_ELIGIBLE_FOR_ORAL_TEST:
             await CQReplySender.ask_yes_no(
